// src/features/dashboard/types/dashboard.types.ts
export interface OrderStats {
  pending: number;
  processing: number;
  completed: number;
  cancelled: number;
  refunded: number;
  failed: number;
  local_facility: number;
  out_for_delivery: number;
}

export interface OrderStatsData {
  today: OrderStats;
  weekly: OrderStats;
  monthly: OrderStats;
  yearly: OrderStats;
}